improve management of actions
Modifications qui seront validées :
	modifié :         src/gestionAction/management.c
	modifié :         src/gestionAction/management.h
	modifié :         src/lib/parserXML/loadXML.c
	modifié :         src/main.c
	modifié :         src/struct/structAction.h

+++ src/struct/structAction.h
@@ -44,6 +44,6 @@ typedef struct
   char* listTimeOut;
 
   char** params;
-  double heureCreation;
+  uint64_t heureCreation;
 }Action;
 #endif //STRUCTACTION_H
